server_v1/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
python-dotenv==1.1.1
gitpython==3.1.45
neo4j==6.0.1
tree-sitter==0.23.2
tree-sitter-language-pack==0.9.1
google-generativeai==0.8.5
google-genai